requirements.txt
discord.py
gspread
oauth2client
flask
requests
pyngrok
